Move message encoders and validators to a separate module
